fix(백준): 오답 수정

+++ BaekJoon/5347.cc
@@ -1,15 +1,13 @@
 #include<bits/stdc++.h>
 using namespace std;
 
-//	백준 문제번호
-//	boj.kr
 
 int gcd(const int& a, const int& b) {
 	if (b == 0)	return a;
 	else return gcd(b, a % b);
 }
 
-int lcm(const int& a, const int& b) {
+long long lcm(const int& a, const int& b) {
 	return (long long)a * b / gcd(a, b);
 }
